Going back to Suwayomi Tracking. Also made sure that MAL won't be available for anybody unless using a release revision >= 1510. Use AniList, seriously.

+++ src/main/java/online/hatsunemiku/tachideskvaadinui/data/tracking/statistics/AniListMangaStatistics.java
@@ -47,12 +47,12 @@ public class AniListMangaStatistics implements MangaStatistics {
   }
 
   @Override
-  public int progress() {
+  public double progress() {
     return progress;
   }
 
   @Override
-  public int score() {
+  public double score() {
     return score;
   }
