+++ src/main/java/online/hatsunemiku/tachideskvaadinui/data/tracking/statistics/MALMangaStatistics.java
@@ -49,12 +49,12 @@ public class MALMangaStatistics implements MangaStatistics {
   }
 
   @Override
-  public int progress() {
+  public double progress() {
     return progress;
   }
 
   @Override
-  public int score() {
+  public double score() {
     return score;
   }
